backend/pyproject.toml
[tool.poetry]
name = "todo-api"
version = "1.0.0"
description = "FastAPI backend for multi-user todo sync"
authors = ["TodoExample Contributors"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
aiosqlite = "^0.19.0"
pydantic = { extras = ["email"], version = "^2.5.3" }
pydantic-settings = "^2.1.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["argon2"], version = "^1.7.4" }
python-multipart = "^0.0.6"
authlib = "^1.3.0"
httpx = "^0.26.0"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
emails = "^0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
faker = "^22.0.0"
black = "^23.12.1"
ruff = "^0.1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
